#377 Refactoring evaluation
