+++ app/Models/HR/Application.php
@@ -7,6 +7,7 @@ use App\Helpers\ContentHelper;
 use App\Models\HR\Applicant;
 use App\Models\HR\ApplicationMeta;
 use App\Models\HR\ApplicationRound;
+use App\Models\HR\Evaluation\ApplicationEvaluation;
 use App\Models\HR\Job;
 use App\User;
 use Illuminate\Database\Eloquent\Model;
