+++ app/Models/HR/Evaluation/Parameter.php
@@ -1,13 +1,12 @@
 <?php
 
-namespace App\Models\HR;
+namespace App\Models\HR\Evaluation;
 
-use App\Models\HR\EvaluationParameterOption;
 use App\Models\HR\Round;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\SoftDeletes;
 
-class EvaluationParameter extends Model
+class Parameter extends Model
 {
     use SoftDeletes;
 
@@ -24,6 +23,6 @@ class EvaluationParameter extends Model
 
     public function options()
     {
-        return $this->hasMany(EvaluationParameterOption::class, 'evaluation_id');
+        return $this->hasMany(ParameterOption::class, 'evaluation_id');
     }
 }
