+++ app/Models/HR/Round.php
@@ -2,7 +2,7 @@
 
 namespace App\Models\HR;
 
-use App\Models\HR\EvaluationParameter;
+use App\Models\HR\Evaluation\Parameter;
 use App\Models\HR\Job;
 use Illuminate\Database\Eloquent\Model;
 
@@ -29,6 +29,6 @@ class Round extends Model
 
     public function evaluationParameters()
     {
-        return $this->belongsToMany(EvaluationParameter::class, 'hr_round_evaluation', 'round_id', 'evaluation_id');
+        return $this->belongsToMany(Parameter::class, 'hr_round_evaluation', 'round_id', 'evaluation_id');
     }
 }
