+++ resources/views/hr/application/application-evaluation.blade.php
@@ -1,0 +1,15 @@
+<div class="d-inline text-danger c-pointer font-weight-bold m-5">Evaluation</div>
+<div class="card border-dark inline-resume" id="sliding_resume" v-cloak v-show="showResumeFrame">
+    <div class="card-header bg-warning">
+        <ul class="nav justify-content-end">
+            <li class="nav-item">
+            </li>
+            <li class="nav-item">
+                <b><span class="c-pointer text-primary" @click="toggleResumeFrame()">Close</span></b>
+            </li>
+        </ul>
+    </div>
+    <div class="card-body">
+        
+    </div>
+</div>
